backend/requirements.txt
fastapi==0.115.13
uvicorn==0.34.3
pandas==2.3.0
openpyxl==3.1.5
python-dotenv==1.1.0
psycopg2-binary==2.9.10
sqlalchemy==2.0.41
pinecone==7.1.0
python-multipart==0.0.18
supabase==2.11.0
pydantic==2.9.2
google-generativeai==0.8.3
openai>=1.0.0